Process.exitcode needs to be an Int64 to cover the range typemin(Int32):typemax(UInt32) that may be returned by spawn on windows (fixes #12176)

+++ base/stream.jl
@@ -1015,7 +1015,7 @@ end
 ## Libuv error handling ##
 type UVError <: Exception
     prefix::AbstractString
-    code::Int64 # TODO: change to Int32 when updating to libuv 1.0
+    code::Int32
     UVError(p::AbstractString,code::Integer)=new(p,code)
 end
